fix check to exit

+++ usr/bin/archboot-aarch64-server-release.sh
@@ -66,11 +66,11 @@ cd "${_ISODIR}" || exit 1
 "archboot-${_ARCH}-release.sh" "${_BUILDDIR}" || rm -r "${_BUILDDIR}"
 # set user rights on files
 # set user rights on files
-[[ -d "${_BUILDDIR}" ]] && exit 1
+[[ -d "${_BUILDDIR}" ]] || exit 1
 chmod 755 "${_BUILDDIR}"
 chown -R "${_USER}" "${_BUILDDIR}"
 chgrp -R "${_GROUP}" "${_BUILDDIR}"
-cd "${_BUILDDIR}" || exit 1
+cd "${_BUILDDIR}"
 # remove sha256sum and install image
 rm sha256sum.txt
 # sign files and create new sha256sum.txt
